+++ support/client-unity/Packages/com.fouridstudio.mizugo-client-unity/Runtime/Exceptions/PacketNullException.cs
@@ -1,0 +1,13 @@
+using System;
+
+namespace Mizugo
+{
+    /// <summary>
+    /// 封包為空異常
+    /// </summary>
+    public class PacketNullException : Exception
+    {
+        public PacketNullException(string name)
+            : base(name + " packet null") { }
+    }
+}