+++ support/client-unity/Packages/com.fouridstudio.mizugo-client-unity/Runtime/Procs/ProtoProc.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.Security.Cryptography;
-using System.Text;
+using System;
 using Google.Protobuf;
 using Google.Protobuf.WellKnownTypes;
 
@@ -12,50 +10,31 @@ namespace Mizugo
     using MessageID = Int32;
 
     /// <summary>
-    /// proto處理器, 封包結構使用ProtoMsg, 可以選擇是否啟用base64編碼或是des-cbc加密
+    /// proto處理器, 封包結構使用ProtoMsg
     /// 訊息定義: support/proto/mizugo/protomsg.proto
-    /// 封包編碼: protobuf編碼成位元陣列, (可選)des-cbc加密, (可選)base64編碼
-    /// 封包解碼: (可選)base64解碼, (可選)des-cbc解密, protobuf解碼成訊息結構
     /// </summary>
-    public partial class ProtoProc : Procmgr
+    public partial class ProtoProc : Procmgr, ICodec
     {
-        /// <summary>
-        /// proto使用的填充模式
-        /// </summary>
-        private const PaddingMode padding = PaddingMode.PKCS7;
-
-        public override byte[] Encode(object input)
+        public object Encode(object input)
         {
             if (input == null)
                 throw new ArgumentNullException("input");
 
-            if (input is not ProtoMsg message)
-                throw new InvalidMessageException("encode");
+            if (input is not ProtoMsg temp)
+                throw new ArgumentException("input");
 
-            var output = message.ToByteArray();
-
-            if (desCBC)
-                output = DesCBC.Encrypt(padding, desKey, desIV, output);
-
-            if (base64)
-                output = Base64.Encode(output);
-
-            return output;
+            return temp.ToByteArray();
         }
 
-        public override object Decode(byte[] input)
+        public object Decode(object input)
         {
             if (input == null)
                 throw new ArgumentNullException("input");
 
-            if (base64)
-                input = Base64.Decode(input);
+            if (input is not byte[] temp)
+                throw new ArgumentException("input");
 
-            if (desCBC)
-                input = DesCBC.Decrypt(padding, desKey, desIV, input);
-
-            var message = ProtoMsg.Parser.ParseFrom(input);
-            return message;
+            return ProtoMsg.Parser.ParseFrom(temp);
         }
 
         public override void Process(object input)
@@ -64,7 +43,7 @@ namespace Mizugo
                 throw new ArgumentNullException("input");
 
             if (input is not ProtoMsg message)
-                throw new InvalidMessageException("process");
+                throw new ArgumentException("input");
 
             var process = Get(message.MessageID);
 
@@ -73,40 +52,6 @@ namespace Mizugo
 
             process(message);
         }
-
-        public ProtoProc SetBase64(bool enable)
-        {
-            base64 = enable;
-            return this;
-        }
-
-        public ProtoProc SetDesCBC(bool enable, string key, string iv)
-        {
-            desCBC = enable;
-            desKey = Encoding.UTF8.GetBytes(key);
-            desIV = Encoding.UTF8.GetBytes(iv);
-            return this;
-        }
-
-        /// <summary>
-        /// 是否啟用base64
-        /// </summary>
-        private bool base64 = false;
-
-        /// <summary>
-        /// 是否啟用des-cbc加密
-        /// </summary>
-        private bool desCBC = false;
-
-        /// <summary>
-        /// des密鑰
-        /// </summary>
-        private byte[] desKey = null;
-
-        /// <summary>
-        /// des初始向量
-        /// </summary>
-        private byte[] desIV = null;
     }
 
     public partial class ProtoProc
@@ -139,7 +84,7 @@ namespace Mizugo
                 throw new ArgumentNullException("input");
 
             if (input is not ProtoMsg data)
-                throw new InvalidMessageException("unmarshal");
+                throw new ArgumentException("input");
 
             messageID = data.MessageID;
             message = data.Message.Unpack<T>();